requirements.txt
# LLM & Agent Framework
groq>=0.9.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
langchain>=0.3.0

# PubMedBERT / NER  (downloads model ~400MB on first run)
transformers>=4.40.0
torch>=2.0.0

# Vector Database
chromadb>=0.5.0

# File format support
openpyxl>=3.1.0          # XLSX read/write
pdfplumber>=0.10.0       # PDF text extraction (preferred)
PyPDF2>=3.0.0          # PDF fallback (install if pdfplumber fails)
python-docx>=1.1.0       # DOCX/DOC read

# Data & Config
pydantic>=2.0.0
python-dotenv>=1.0.0

# UI
rich>=13.0.0
flask>=3.0.0
